pyproject.toml
[project]
name = "gen-ai-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.1",
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.10",
    "langchain-astradb>=1.0.0",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.10",
    "langchain-pinecone>=0.2.13",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pypdf>=6.7.3",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.54.0",
    "tiktoken>=0.12.0",
    "unstructured[all-docs]>=0.21.5",
    "wikipedia>=1.4.0",
]
